S2EExecutor: moved function handlers to a separate file
